Basics views of the application
The Accounts, Forecast and Transaction lists are in place.
The ability to add and delete accounts and for them to persist using Core Data is in place.
Transactions nest underneath each forecasted month (currently displaying all transaction in every month).
The ability to add and delete transactions (persisting not yet supported).